requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
sqlalchemy==2.0.23
pydantic==2.5.2
google-generativeai==0.3.2
requests==2.31.0
scipy==1.11.4
shap==0.43.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0
